Replace uses of the "mimetype" arg of HttpResponse with "content_type".
mimetype has been deprecated for a long time and has been removed from
django in the development version. This commit uses content_type for
the default value but allows specifying either, regardless of the
version of Django. Specifying "mimetype" will work for users of
django-oauth2-provider if their version of django supports it (the
deprecated argument has precedence), and will break if it doesn't.