Handle module name changes for Django >=1.5
To allow Python 3 compatibility, Django 1.5 renamed smart_unicode in
django.utils.encoding as smart_text (as Python 3's str type is unicode
by default). To accomodate running on earlier Django versions, we do an
import of the old name; if we fail, we import smart_text as
smart_unicode instead (with the assumption that Django is actually in
the path).