Fix up usage of urlparse to import via Six.
urlparse in Python 3 has been subsumed into urllib; to support both
Python 2 and Python 3, we import urlparse via Six.

This commit also fixes a bug in the tests. There is an incorrect
assumption that parse_qs parses just the query string of a full URL,
when given a full URL; it doesn't, and instead parse_qs assumes the
caller passes in just the query string part of the URL. When given a
full URL, the rest of the URL including the first attribute in the query
string is assigned as the first key of the resulting dictionary returned
by parse_qs. Because the 'code' key incidentally ends up being the
second attribute in the query string, the test passes. However, because
there was a change in the ordering of the query string, no 'code' key
could be found when running in Python 3.