Fix bytes vs str usage.
Because Python 3 is strict when it comes to byte strings vs. text
strings, and because Python 3 does not do any automatic coercion between
the two types, we need to be more explicit.

In particular, when using response.content (which is in really a byte
string), we need to convert it to a particular character encoding before
treating it as a text string. We assume UTF-8 for the tests,
which *will* break in other encodings, but since these are the tests, we
don't mind assuming UTF-8.

As well, because encode() in the str type in Python 3 no longer supports
non-character encodings such as Base64, we need to use codecs to do the
conversion.

We also fix the assumption that shortuuid.uuid() returns a byte
string (which it doesn't) when passing to hashlib.