Don't iterate through the session keys while modifying the session.
This breaks in Python 3, as modifying a dictionary while iterating
through its keys raises an exception, because dict.keys() now returns a
view instead of a materialized list. We instead pull out the session
keys as a list and iterate through that instead, which works for both
Python 3 and Python 2 (although in Python 2 this creates an additional
list).