Refactor AccessToken view to allow adding extra parameters.
OAuth2 allows for extra parameters in the access token response, which
is frequently done by OAuth2 providers to fit their needs.

It is also required for protocols such as OpenID Connect, build
on top of OAuth2, which adds a id_token parameter.

Subclesses of AccessToken, should be able to easily add more parameters,
without repeating code of the parent class.